unshare: move to pkg/
