+++ cmd/buildah/main.go
@@ -5,7 +5,7 @@ import (
 	"github.com/containers/buildah"
 	"os"
 
-	"github.com/containers/buildah/unshare"
+	"github.com/containers/buildah/pkg/unshare"
 	"github.com/containers/storage"
 	ispecs "github.com/opencontainers/image-spec/specs-go"
 	rspecs "github.com/opencontainers/runtime-spec/specs-go"
